+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2025-01-24
+
+### Collection
+
+- *boot_image* module plugin added.
+- *directory_info* module plugin added.
+
+### *directory* module plugin
+
+- Replaced `name` and `parent_directory` options with `path` option.
+- Renamed `mdt_directory_path` option to `mdt_share_path`.
+
+### *import_drivers* module plugin
+
+- Renamed `mdt_directory_path` option to `mdt_share_path`.
+
 ## [1.0.0] - 2025-01-23
 
 ### Collection